requirements.txt
ultralytics
streamlit
torch
opencv-python
numpy
Pillow
pandas
